Added setup.py to install seam_erasure
* Renamed package SeamErasure to seam_erasure
