+++ .gitignore
@@ -158,3 +158,4 @@ Temporary Items
 ###############################################################################
 # Mine
 travis-results.png
+docs
