--- SeamErasure/__init__.py
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-"""SeamErasure: Seamlessly Erase Seams."""
-
-# from . import seam_erasure
-# from . import obj_reader
-# from . import util
-#
-# __all__ = ["seam_erasure", "obj_reader", "util"]
-#
-# __version__ = "0.0.1-alpha.0"
-# __author__ = "Zachary Ferguson"
-# __email__ = "zfergus@nyu.edu"
