+++ seam_erasure/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+"""seam_erasure: Seamlessly erase seams from your favorite 3D models."""
+
+__author__ = "Zachary Ferguson"
+__copyright__ = "Copyright 2016, Zachary Ferguson"
+__credits__ = "Zachary Ferguson, Yotam Gingold, Songrun Liu, Alec Jacobson"
+__license__ = "MIT"
+__version__ = "0.1.0"
+__maintainer__ = "Zachary Ferguson"
+__email__ = "zfergus@nyu.edu"
+__status__ = "Production"
